+++ src/amplitude-client.js
@@ -19,6 +19,7 @@ import getHost from './get-host';
 import baseCookie from './base-cookie';
 import { AmplitudeServerZone, getEventLogApi } from './server-zone';
 import ConfigManager from './config-manager';
+import GlobalScope from './global-scope';
 
 import { AmplitudeCore } from '@amplitude/amplitude-core';
 
@@ -30,7 +31,7 @@ import { AmplitudeCore } from '@amplitude/amplitude-core';
  * @example var amplitudeClient = new AmplitudeClient();
  */
 var AmplitudeClient = function AmplitudeClient(instanceName) {
-  if (!isBrowserEnv()) {
+  if (!isBrowserEnv() && !utils.isWebWorkerEnvironment()) {
     utils.log.warn(
       'amplitude-js will not work in a non-browser environment. If you are planning to add Amplitude to a node environment, please use @amplitude/node',
     );
@@ -88,7 +89,11 @@ AmplitudeClient.prototype.init = function init(apiKey, opt_userId, opt_config, o
     this._core = AmplitudeCore.getInstance(this._instanceName);
 
     _parseConfig(this.options, opt_config);
-    if (isBrowserEnv() && window.Prototype !== undefined && Array.prototype.toJSON) {
+    if (
+      (isBrowserEnv() || utils.isWebWorkerEnvironment()) &&
+      GlobalScope.Prototype !== undefined &&
+      Array.prototype.toJSON
+    ) {
       prototypeJsFix();
       utils.log.warn(
         'Prototype.js injected Array.prototype.toJSON. Deleting Array.prototype.toJSON to prevent double-stringify',
@@ -251,7 +256,7 @@ AmplitudeClient.prototype.init = function init(apiKey, opt_userId, opt_config, o
         // Monitoring just page exits because that is the most requested feature for now
         // "If you're specifically trying to detect page unload events, the pagehide event is the best option."
         // https://developer.mozilla.org/en-US/docs/Web/API/Window/pagehide_event
-        window.addEventListener(
+        GlobalScope.addEventListener(
           'pagehide',
           () => {
             handleVisibilityChange();
@@ -277,7 +282,7 @@ AmplitudeClient.prototype.init = function init(apiKey, opt_userId, opt_config, o
     editor.commit();
   } catch (err) {
     utils.log.error(err);
-    if (type(opt_config.onError) === 'function') {
+    if (opt_config && type(opt_config.onError) === 'function') {
       opt_config.onError(err);
     }
   }
@@ -295,7 +300,7 @@ AmplitudeClient.prototype.deleteLowerLevelDomainCookies = function () {
   const cookieHost =
     this.options.domain && this.options.domain[0] === '.' ? this.options.domain.slice(1) : this.options.domain;
 
-  if (!cookieHost) {
+  if (!cookieHost || !utils.isWebWorkerEnvironment()) {
     return;
   }
 
@@ -774,14 +779,14 @@ var _sendParamsReferrerUserProperties = function _sendParamsReferrerUserProperti
  * @private
  */
 AmplitudeClient.prototype._getReferrer = function _getReferrer() {
-  return document.referrer;
+  return typeof document !== 'undefined' ? document.referrer : '';
 };
 
 /**
  * @private
  */
 AmplitudeClient.prototype._getUrlParams = function _getUrlParams() {
-  return location.search;
+  return GlobalScope.location.search;
 };
 
 /**
@@ -1824,7 +1829,7 @@ AmplitudeClient.prototype.sendEvents = function sendEvents() {
     }
     this._sending = true;
   }
-  var protocol = this.options.forceHttps ? 'https' : 'https:' === window.location.protocol ? 'https' : 'http';
+  var protocol = this.options.forceHttps ? 'https' : 'https:' === GlobalScope.location.protocol ? 'https' : 'http';
   var url = protocol + '://' + this.options.apiEndpoint;
 
   // fetch events to send