+++ src/get-host.js
@@ -1,7 +1,18 @@
+import GlobalScope from './global-scope';
+
 const getHost = (url) => {
-  const a = document.createElement('a');
-  a.href = url;
-  return a.hostname || location.hostname;
+  if (url) {
+    if (typeof document !== 'undefined') {
+      const a = document.createElement('a');
+      a.href = url;
+      return a.hostname || GlobalScope.location.hostname;
+    }
+    if (typeof URL === 'function') {
+      const u = new URL(url);
+      return u.hostname || GlobalScope.location.hostname;
+    }
+  }
+  return GlobalScope.location.hostname;
 };
 
 export default getHost;