+++ src/global-scope.js
@@ -1,0 +1,17 @@
+/* global globalThis */
+const GlobalScope = (() => {
+  if (typeof globalThis !== 'undefined') {
+    return globalThis;
+  }
+  if (typeof window !== 'undefined') {
+    return window;
+  }
+  if (typeof self !== 'undefined') {
+    return self;
+  }
+  if (typeof global !== 'undefined') {
+    return global;
+  }
+})();
+
+export default GlobalScope;