+++ src/top-domain.js
@@ -1,6 +1,7 @@
 import baseCookie from './base-cookie';
 import base64Id from './base64Id';
 import getHost from './get-host';
+import utils from './utils';
 
 // Utility that finds top level domain to write to
 const topDomain = (url) => {
@@ -8,6 +9,8 @@ const topDomain = (url) => {
   const parts = host.split('.');
   const levels = [];
   const cname = '_tldtest_' + base64Id();
+
+  if (utils.isWebWorkerEnvironment()) return '';
 
   for (let i = parts.length - 2; i >= 0; --i) {
     levels.push(parts.slice(i).join('.'));