+++ CHANGELOG.md
@@ -1,3 +1,25 @@
+# [8.15.0](https://github.com/amplitude/amplitude-javascript/compare/v8.14.1...v8.15.0) (2022-01-07)
+
+
+### Features
+
+* make snippet public in NPM ([#478](https://github.com/amplitude/amplitude-javascript/issues/478)) ([8f512f0](https://github.com/amplitude/amplitude-javascript/commit/8f512f0a3811acaffd3ad7662dcf0ca6738e0522))
+
+## [8.14.1](https://github.com/amplitude/amplitude-javascript/compare/v8.14.0...v8.14.1) (2021-12-23)
+
+
+### Bug Fixes
+
+* global-scope reference error in nodejs ([#474](https://github.com/amplitude/amplitude-javascript/issues/474)) ([bdce39d](https://github.com/amplitude/amplitude-javascript/commit/bdce39daedf8ff965e43f038cd2c73319c63930d))
+
+# [8.14.0](https://github.com/amplitude/amplitude-javascript/compare/v8.13.1...v8.14.0) (2021-12-21)
+
+
+### Features
+
+* add runNewSessionStartCallback on new sessionId on log event ([#469](https://github.com/amplitude/amplitude-javascript/issues/469)) ([bb8b26b](https://github.com/amplitude/amplitude-javascript/commit/bb8b26b23267ed6fd5adc393f873e2d1103f4030))
+* support web worker env ([#467](https://github.com/amplitude/amplitude-javascript/issues/467)) ([52abaf0](https://github.com/amplitude/amplitude-javascript/commit/52abaf062492217ddf8d556b85509ce00e006f25))
+
 ## [8.13.1](https://github.com/amplitude/amplitude-javascript/compare/v8.13.0...v8.13.1) (2021-12-03)